perf(page-load): code-split teacher shell + lazy-init deep-link import hooks (#2041)
* perf(page-load): code-split Dock + Sidebar out of synchronous teacher mount

The board canvas now paints before the Dock/Sidebar chunks load; both stream
in on the next microtask behind a Suspense boundary (fallback reserves the
dock footprint only when a dock will actually render, so annotation/read-only
boards show no skeleton). Both are position:fixed overlays, so deferring them
a tick cannot shift the board's layout.

DashboardView keeps the global keydown listener and the useGesture surface, so
keyboard shortcuts and pan/zoom are unaffected by the shell mounting a tick
later.

Extends the page-load harness to record firstCommitMedianMs (time-to-first-
paint) alongside the existing settle-window sum. The first-commit duration for
DashboardView drops ~16ms -> ~6.9ms (medians of 6 iterations x3 runs); the
settle-window sum is unchanged in a real browser (same total work, just
deferred past first paint).


* perf(page-load): lazy-init DashboardView deep-link import hooks

DashboardView called five Firestore-listener hooks at the top of every
teacher mount — useQuiz, useQuizAssignments, useVideoActivity,
useVideoActivityAssignments, usePlcs — but consumed ONLY their import/peek
callbacks (never their assignments/quizzes/plcs lists; the Quiz and
VideoActivity widgets run their own independent listeners). So every normal
teacher load paid for five live onSnapshot subscriptions plus the associated
callback/effect setup to service a code path that only matters when a share
link is actually being imported.

Extracts that machinery (the five hooks, the open*WidgetToTab helpers, the
three pending-share effects, runAssignmentImport / peekAndDispatchImport /
runVideoActivityImport, and the import mode-picker modal) into a new
DeepLinkShareImporter, mounted lazily and only once a pending share id
appears. The mount is latched (imports clear their pending id synchronously,
so a bare presence gate would unmount mid-import) and behaviour is identical:
the effects fire on mount with the same pending id, clear it synchronously to
avoid the triple-import race, and surface the same toasts / picker.

Net effect on the common teacher load: zero of these Firestore listeners are
opened (a real cost reduction not visible to the mocked perf harness), and the
import-machinery callbacks/effects no longer run at mount.

No new Firestore/Storage/AI usage — strictly fewer listeners on the common
path; the import paths issue exactly the same reads/writes as before.


* test(page-load): await lazy Sidebar/Dock; refresh page-load baseline

DashboardView's Sidebar and Dock are now code-split (React.lazy), so the
'renders correctly' test must await their Suspense resolution via
findByTestId instead of querying synchronously.

Regenerates page-load-baseline.json to include the new firstCommitMedianMs
metric and DashboardView's post-split profile (commits 1 -> 2 as the dock/
sidebar now land in a second commit). Per-page actualDuration values are
machine-dependent and indicative only.


* perf(page-load): fix committed baseline + reviewer nits

- page-load-baseline.json: regenerate with the OPTIMIZED DashboardView. The
  prior commit accidentally captured the baseline snapshot (a temporary
  measurement run had overwritten the file), showing commits=1 / 18.4ms. The
  corrected artifact shows the post-split profile (commits=2, firstCommit
  ~6.7ms), matching the PR's claims.
- DeepLinkShareImporter: retag the three logError context strings from
  'DashboardView.*' to 'DeepLinkShareImporter.*' so error monitoring routes
  to the component that actually owns the code.
- Document the one-time plcsLoading latency shift on URL-parameter deep links
  (the usePlcs listener now starts with the importer rather than at teacher
  mount); behaviour is unchanged, only a small deep-link-path latency trade.


* perf(page-load): fix stale-closure in importer cancel-path Retry

The QuizAssignmentImportModeModal onClose cancel path created its Retry toast
with a direct `peekAndDispatchImport` reference, pinning the dispatcher (and
its `plcs` membership snapshot) captured when the toast was created. The
catch-path Retry already routes through `peekAndDispatchImportRef.current`;
match it here so a Retry clicked long after cancel runs against the freshest
PLC membership snapshot. Behaviour-identical on the common path; only affects
the rare cancel-then-retry-after-plcs-changed window.


---------
